"Updated navbar and homepage styles, added SignUp-btn class"
This commit includes changes to the NavBar.css and HomePage.css files. In NavBar.css, there are adjustments to margins, addition of a SignUp-btn class, and changes to hover and active styles for buttons. In HomePage.css, there is a change in the background color of the signup-btn-2 class. Additionally, a button with the SignUp-btn class has been added to Navbar.js.